Mark TOMEE as an allowable failure
Currently, TOMEE_MANAGED_1.5 fails on every build, presumably due to configuration issues.  This change marks that as allowable, so if the other four builds succeed, it will show green rather than red as it does now.  The TOMEE build will still show red on the details screen, but the summary view on GitHub will show green.  

This way, if someone fixes TOMEE, it will be visible in Travis.  But this way, red means failed rather than "check if anything other than TOMEE failed."  A typical build with all successes except a TOMEE failure will show as green.  At least that's what https://docs.travis-ci.com/user/customizing-the-build#rows-that-are-allowed-to-fail says.  I can't test without committing and requesting a pull.